Enemies updates and tweaks
Added multiple enemies to the scene, made them not a trigger so they can collide with each other and not overlap.

Tweaked some settings so the enemy can catch up with the player if they are walking but if the player is running they can out run/manoeuvere them.

Changed the players health to be out of 100 so that it can translate to a bar better.